requirements.txt
pyTelegramBotAPI
psutil
python-dotenv
pySMART
